topdown: Add http.send support for setting the TLS server name
If the caller sets the Host header, then we can reasonably impute that
they wanted to use that at the TLS layer as well. To handle test cases
where the caller needs to explicitly set the TLS server name, add a
`tls_server_name` parameter that overrides both the `Host` header and
the Go default.

Signed-off-by: James Peach <jpeach@vmware.com>